extraction of buffers in client

+++ src/webglpp.cpp
@@ -15,21 +15,18 @@ WebGLpp::send() {
     // retrieve the data
     nlohmann::json jbuffer;
     if (buffer.type() == typeid(float).name()) {
-
-      // add the buffer data to the json
       std::vector<float> x = buffer.Float32Array();
       jbuffer["data"] = x;
       jbuffer["type"] = "Float32Array";
     }
-    else if (buffer.type() == typeid(unsigned int).name()) {
-      std::vector<unsigned int> x = buffer.Uint16Array();
+    else if (buffer.type() == typeid(unsigned short).name()) {
+      std::vector<unsigned short> x = buffer.Uint16Array();
       jbuffer["data"] = x;
       jbuffer["type"] = "Uint16Array";
     }
     jbuffer["tag"]    = buffer.tag();
     jbuffer["index"]  = i;
     jbuffer["target"] = target_name(buffer.target());
-
 
     std::cout << jbuffer.dump() << std::endl;
